# flaskserver/flask_server/swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: OnTime Attendance Tracking API
  description: "API documentation for OnTime attendance tracking software.This API\
    \ documentation highlights the various requests we would implement using the rest\
    \ API . They are generally shown through put, get and post requests. The attendance\
    \ tracking in general requires database changes on a major amount and hence need\
    \ to be handled well and hence we tested our API documentation using swagger."
  version: 1.0.0
servers:
- url: http://ontime.com/api
paths:
  /member_login:
    post:
      tags:
      - Login
      summary: Login for a member of an organization
      description: Login Option for members
      operationId: member_login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/member_login_body'
      responses:
        "200":
          description: successful member login
        "400":
          description: Invalid Login
      x-openapi-router-controller: swagger_server.controllers.login_controller
...
  /admin_login:
    post:
      tags:
      - Login
      summary: Login for admin of an organization
      description: This is for login of a whole organization which has only one admin
        per one organization.
      operationId: admin_login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/admin_login_body'
      responses:
        "200":
          description: Successful response
        "400":
          description: Invalid username/password supplied
      x-openapi-router-controller: swagger_server.controllers.login_controller
...
components:
  schemas:
...
    member_login_body:
      required:
      - member_id
      - organization_id
      - password
      type: object
      properties:
        organization_id:
          type: string
          example: CompanyXYZ
        member_id:
          type: string
          example: NMember87
        password:
          type: string
          format: password
          example: PussInBoots69
...
    admin_login_body:
      required:
      - organization_id
      - password
      type: object
      properties:
        organization_id:
          type: string
          example: CompanyXYZ
        password:
          type: string
          format: password
          example: PussInBoots69
...